Split Multigas collector interface from implementation
