+++ execution/multigascollector/file_collector_test.go
@@ -17,83 +17,6 @@ import (
 	"github.com/offchainlabs/nitro/execution/multigascollector/proto"
 )
 
-func TestTransactionMultiGasToProto(t *testing.T) {
-	tests := []struct {
-		name     string
-		tx       *TransactionMultiGas
-		expected func(*testing.T, *proto.TransactionMultiGasData)
-	}{
-		{
-			name: "transaction with all gas dimensions and optional fields",
-			tx: &TransactionMultiGas{
-				TxHash:  []byte{0x12, 0x34, 0x56},
-				TxIndex: 0,
-				MultiGas: *multigas.MultiGasFromMap(map[multigas.ResourceKind]uint64{
-					multigas.ResourceKindComputation:   100,
-					multigas.ResourceKindHistoryGrowth: 50,
-					multigas.ResourceKindStorageAccess: 200,
-					multigas.ResourceKindStorageGrowth: 1000,
-					multigas.ResourceKindUnknown:       10,
-				}),
-			},
-			expected: func(t *testing.T, proto *proto.TransactionMultiGasData) {
-				assert.Equal(t, []byte{0x12, 0x34, 0x56}, proto.TxHash)
-				assert.Equal(t, uint32(0), proto.TxIndex)
-				assert.Equal(t, uint64(100), proto.MultiGas.Computation)
-				assert.Equal(t, uint64(50), proto.MultiGas.HistoryGrowth)
-				assert.Equal(t, uint64(200), proto.MultiGas.StorageAccess)
-				assert.Equal(t, uint64(1000), proto.MultiGas.StorageGrowth)
-				assert.NotNil(t, proto.MultiGas.Unknown)
-				assert.Equal(t, uint64(10), *proto.MultiGas.Unknown)
-				assert.Nil(t, proto.MultiGas.Refund) // No refund in test data
-			},
-		},
-		{
-			name: "transaction with minimal gas dimensions (no optional fields)",
-			tx: &TransactionMultiGas{
-				TxHash:  []byte{0x78, 0x9a, 0xbc},
-				TxIndex: 1,
-				MultiGas: *multigas.MultiGasFromMap(map[multigas.ResourceKind]uint64{
-					multigas.ResourceKindComputation: 150,
-				}),
-			},
-			expected: func(t *testing.T, proto *proto.TransactionMultiGasData) {
-				assert.Equal(t, []byte{0x78, 0x9a, 0xbc}, proto.TxHash)
-				assert.Equal(t, uint32(1), proto.TxIndex)
-				assert.Equal(t, uint64(150), proto.MultiGas.Computation)
-				assert.Equal(t, uint64(0), proto.MultiGas.HistoryGrowth)
-				assert.Equal(t, uint64(0), proto.MultiGas.StorageAccess)
-				assert.Equal(t, uint64(0), proto.MultiGas.StorageGrowth)
-				assert.Nil(t, proto.MultiGas.Unknown) // Should be nil since value was 0
-				assert.Nil(t, proto.MultiGas.Refund)  // Should be nil since value was 0
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			protoData := tt.tx.ToProto()
-			tt.expected(t, protoData)
-		})
-	}
-}
-
-func TestBlockInfoToProto(t *testing.T) {
-	blockInfo := &BlockInfo{
-		BlockNumber:    12345,
-		BlockHash:      []byte{0xab, 0xcd, 0xef},
-		BlockTimestamp: 1234567890,
-	}
-
-	protoData := blockInfo.ToProto()
-
-	// Verify block metadata
-	assert.Equal(t, blockInfo.BlockNumber, protoData.BlockNumber)
-	assert.Equal(t, blockInfo.BlockHash, protoData.BlockHash)
-	assert.Equal(t, blockInfo.BlockTimestamp, protoData.BlockTimestamp)
-	assert.Empty(t, protoData.Transactions) // No transactions initially
-}
-
 func TestIdleCollector(t *testing.T) {
 	tests := []struct {
 		name      string
@@ -130,7 +53,7 @@ func TestIdleCollector(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			ctx := context.Background()
 
-			collector, err := NewCollector(tt.config)
+			collector, err := NewFileCollector(tt.config)
 
 			if tt.expectErr != nil {
 				assert.Error(t, err)
@@ -153,7 +76,7 @@ func TestDataCollection(t *testing.T) {
 	testCases := []struct {
 		name        string
 		batchSize   uint64
-		inputData   []*CollectorMessage
+		inputData   []*Message
 		expectFiles int
 	}{
 		{
@@ -165,10 +88,15 @@ func TestDataCollection(t *testing.T) {
 		{
 			name:      "empty block",
 			batchSize: 10,
-			inputData: []*CollectorMessage{
-				{Type: CollectorMsgStartBlock},
+			inputData: []*Message{
 				{
-					Type: CollectorMsgFinaliseBlock,
+					Type: MsgStartBlock,
+					Block: &BlockInfo{
+						BlockNumber: 20001,
+					},
+				},
+				{
+					Type: MsgFinaliseBlock,
 					Block: &BlockInfo{
 						BlockNumber:    20001,
 						BlockHash:      []byte{0xaa, 0xbb, 0xcc},
@@ -181,9 +109,9 @@ func TestDataCollection(t *testing.T) {
 		{
 			name:      "discarded transaction",
 			batchSize: 10,
-			inputData: []*CollectorMessage{
+			inputData: []*Message{
 				{
-					Type: CollectorMsgTransaction,
+					Type: MsgTransaction,
 					Transaction: &TransactionMultiGas{
 						TxHash:  []byte{0x12, 0x34, 0x56},
 						TxIndex: 0,
@@ -196,17 +124,27 @@ func TestDataCollection(t *testing.T) {
 						}),
 					},
 				},
-				{Type: CollectorMsgStartBlock},
+				{
+					Type: MsgStartBlock,
+					Block: &BlockInfo{
+						BlockNumber: 20001,
+					},
+				},
 			},
 			expectFiles: 0,
 		},
 		{
 			name:      "single block - one transaction",
 			batchSize: 1,
-			inputData: []*CollectorMessage{
-				{Type: CollectorMsgStartBlock},
+			inputData: []*Message{
 				{
-					Type: CollectorMsgTransaction,
+					Type: MsgStartBlock,
+					Block: &BlockInfo{
+						BlockNumber: 12345,
+					},
+				},
+				{
+					Type: MsgTransaction,
 					Transaction: &TransactionMultiGas{
 						TxHash:  []byte{0x12, 0x34, 0x56},
 						TxIndex: 0,
@@ -220,7 +158,7 @@ func TestDataCollection(t *testing.T) {
 					},
 				},
 				{
-					Type: CollectorMsgFinaliseBlock,
+					Type: MsgFinaliseBlock,
 					Block: &BlockInfo{
 						BlockNumber:    12345,
 						BlockHash:      []byte{0xab, 0xcd, 0xef},
@@ -233,10 +171,15 @@ func TestDataCollection(t *testing.T) {
 		{
 			name:      "start new block without finalising previous -> drop unfinalised txs",
 			batchSize: 10,
-			inputData: []*CollectorMessage{
-				{Type: CollectorMsgStartBlock},
+			inputData: []*Message{
 				{
-					Type: CollectorMsgTransaction,
+					Type: MsgStartBlock,
+					Block: &BlockInfo{
+						BlockNumber: 30000,
+					},
+				},
+				{
+					Type: MsgTransaction,
 					Transaction: &TransactionMultiGas{
 						TxHash:  []byte{0x01},
 						TxIndex: 0,
@@ -246,9 +189,14 @@ func TestDataCollection(t *testing.T) {
 					},
 				},
 				// Start a new block before finalising the previous -> prior tx is dropped
-				{Type: CollectorMsgStartBlock},
 				{
-					Type: CollectorMsgTransaction,
+					Type: MsgStartBlock,
+					Block: &BlockInfo{
+						BlockNumber: 30001,
+					},
+				},
+				{
+					Type: MsgTransaction,
 					Transaction: &TransactionMultiGas{
 						TxHash:  []byte{0x02},
 						TxIndex: 0,
@@ -258,7 +206,7 @@ func TestDataCollection(t *testing.T) {
 					},
 				},
 				{
-					Type: CollectorMsgFinaliseBlock,
+					Type: MsgFinaliseBlock,
 					Block: &BlockInfo{
 						BlockNumber:    30001,
 						BlockHash:      []byte{0x01, 0x02, 0x03},
@@ -271,10 +219,15 @@ func TestDataCollection(t *testing.T) {
 		{
 			name:      "multiple blocks - single batch",
 			batchSize: 3,
-			inputData: []*CollectorMessage{
-				{Type: CollectorMsgStartBlock},
+			inputData: []*Message{
 				{
-					Type: CollectorMsgTransaction,
+					Type: MsgStartBlock,
+					Block: &BlockInfo{
+						BlockNumber: 12345,
+					},
+				},
+				{
+					Type: MsgTransaction,
 					Transaction: &TransactionMultiGas{
 						TxHash:  []byte{0x12, 0x34, 0x56},
 						TxIndex: 0,
@@ -286,16 +239,21 @@ func TestDataCollection(t *testing.T) {
 					},
 				},
 				{
-					Type: CollectorMsgFinaliseBlock,
+					Type: MsgFinaliseBlock,
 					Block: &BlockInfo{
 						BlockNumber:    12345,
 						BlockHash:      []byte{0xab, 0xcd, 0xef},
 						BlockTimestamp: 1234567890,
 					},
 				},
-				{Type: CollectorMsgStartBlock},
 				{
-					Type: CollectorMsgTransaction,
+					Type: MsgStartBlock,
+					Block: &BlockInfo{
+						BlockNumber: 12346,
+					},
+				},
+				{
+					Type: MsgTransaction,
 					Transaction: &TransactionMultiGas{
 						TxHash:  []byte{0x45, 0x67, 0x89},
 						TxIndex: 1,
@@ -307,7 +265,7 @@ func TestDataCollection(t *testing.T) {
 					},
 				},
 				{
-					Type: CollectorMsgTransaction,
+					Type: MsgTransaction,
 					Transaction: &TransactionMultiGas{
 						TxHash:  []byte{0x78, 0x9a, 0xbc},
 						TxIndex: 2,
@@ -319,7 +277,7 @@ func TestDataCollection(t *testing.T) {
 					},
 				},
 				{
-					Type: CollectorMsgFinaliseBlock,
+					Type: MsgFinaliseBlock,
 					Block: &BlockInfo{
 						BlockNumber:    12346,
 						BlockHash:      []byte{0xde, 0xf1, 0x23},
@@ -332,10 +290,15 @@ func TestDataCollection(t *testing.T) {
 		{
 			name:      "multiple blocks - multiple batches",
 			batchSize: 2,
-			inputData: []*CollectorMessage{
-				{Type: CollectorMsgStartBlock},
+			inputData: []*Message{
 				{
-					Type: CollectorMsgTransaction,
+					Type: MsgStartBlock,
+					Block: &BlockInfo{
+						BlockNumber: 12345,
+					},
+				},
+				{
+					Type: MsgTransaction,
 					Transaction: &TransactionMultiGas{
 						TxHash:  []byte{0x12, 0x34, 0x56},
 						TxIndex: 0,
@@ -348,16 +311,21 @@ func TestDataCollection(t *testing.T) {
 					},
 				},
 				{
-					Type: CollectorMsgFinaliseBlock,
+					Type: MsgFinaliseBlock,
 					Block: &BlockInfo{
 						BlockNumber:    12345,
 						BlockHash:      []byte{0xab, 0xcd, 0xef},
 						BlockTimestamp: 1234567890,
 					},
 				},
-				{Type: CollectorMsgStartBlock},
 				{
-					Type: CollectorMsgTransaction,
+					Type: MsgStartBlock,
+					Block: &BlockInfo{
+						BlockNumber: 12346,
+					},
+				},
+				{
+					Type: MsgTransaction,
 					Transaction: &TransactionMultiGas{
 						TxHash:  []byte{0x78, 0x9a, 0xbc},
 						TxIndex: 1,
@@ -369,16 +337,21 @@ func TestDataCollection(t *testing.T) {
 					},
 				},
 				{
-					Type: CollectorMsgFinaliseBlock,
+					Type: MsgFinaliseBlock,
 					Block: &BlockInfo{
 						BlockNumber:    12346,
 						BlockHash:      []byte{0xde, 0xf1, 0x23},
 						BlockTimestamp: 1234567891,
 					},
 				},
-				{Type: CollectorMsgStartBlock},
 				{
-					Type: CollectorMsgTransaction,
+					Type: MsgStartBlock,
+					Block: &BlockInfo{
+						BlockNumber: 12347,
+					},
+				},
+				{
+					Type: MsgTransaction,
 					Transaction: &TransactionMultiGas{
 						TxHash:  []byte{0xab, 0xcd, 0xef},
 						TxIndex: 2,
@@ -390,7 +363,7 @@ func TestDataCollection(t *testing.T) {
 					},
 				},
 				{
-					Type: CollectorMsgFinaliseBlock,
+					Type: MsgFinaliseBlock,
 					Block: &BlockInfo{
 						BlockNumber:    12347,
 						BlockHash:      []byte{0x45, 0x67, 0x89},
@@ -411,7 +384,7 @@ func TestDataCollection(t *testing.T) {
 				BatchSize: int(tt.batchSize), //nolint:gosec
 			}
 
-			collector, err := NewCollector(config)
+			collector, err := NewFileCollector(config)
 			require.NoError(t, err)
 			collector.Start(context.Background())
 
@@ -439,11 +412,11 @@ func TestDataCollection(t *testing.T) {
 			var curTxs []*proto.TransactionMultiGasData
 			for _, m := range tt.inputData {
 				switch m.Type {
-				case CollectorMsgStartBlock:
+				case MsgStartBlock:
 					curTxs = nil // drop unfinalised txs
-				case CollectorMsgTransaction:
+				case MsgTransaction:
 					curTxs = append(curTxs, m.Transaction.ToProto())
-				case CollectorMsgFinaliseBlock:
+				case MsgFinaliseBlock:
 					blk := m.Block.ToProto()
 					if len(curTxs) > 0 {
 						blk.Transactions = append(blk.Transactions, curTxs...)
